+++ plugin/src/net/groboclown/idea/p4ic/v2/changes/P4ShelvedContentRevision.java
@@ -20,7 +20,8 @@ import com.intellij.openapi.vcs.VcsException;
 import com.intellij.openapi.vcs.changes.ContentRevision;
 import com.intellij.openapi.vcs.history.VcsRevisionNumber;
 import net.groboclown.idea.p4ic.v2.server.cache.ClientServerRef;
-import net.groboclown.idea.p4ic.v2.server.util.DepotFilePath;
+import net.groboclown.idea.p4ic.v2.server.cache.state.P4ShelvedFile;
+import net.groboclown.idea.p4ic.v2.server.util.ShelvedFilePath;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -28,11 +29,11 @@ public class P4ShelvedContentRevision implements ContentRevision {
     private static final P4CurrentContentRevision.CurrentRevisionNumber
             HAVE_REV = new P4CurrentContentRevision.CurrentRevisionNumber();
 
-    private final DepotFilePath filePath;
+    private final ShelvedFilePath filePath;
 
     public P4ShelvedContentRevision(@Nullable Project project, @NotNull ClientServerRef clientServerRef,
-            @NotNull String depotPath) {
-        this.filePath = new DepotFilePath(project, clientServerRef, depotPath);
+            @NotNull P4ShelvedFile shelvedFile) {
+        this.filePath = new ShelvedFilePath(shelvedFile);
     }
 
     @Nullable